Alpha 0.20.2
Made a better core and view structure + some extra functions.
the files application.php, database.php, login.php, register.php,
index.php and config.php still require php doc.
