+++ application/config/config.php
@@ -1,5 +1,7 @@
 <?php
 
+    /**
+    */
     define("DB_HOST", "127.0.0.1");
     define("DB_NAME", "komodo");
     define("DB_USER", "root");
