--- application/view/hero.php
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<head>
-</head>
-<body>
-    <main class="overall">
-        <p>
-           Hey, <?php echo $_SESSION['user_name']; ?>. You are logged in.<br>
-           Try to close this browser tab and open it again. Still logged in! ;)
-        </p>
-        <h1>HIER MOETEN HERO'S STAAN OFZO<h1>
-        <a href="index.php?logout" >log out</a>
-    </main>
-</body>
