+++ application/view/indexes/game.php
@@ -1,0 +1,1 @@
+<script>window.data = <?=$komodo->getRequiredGameDat()?></script>
