+++ application/view/indexes/hero.php
@@ -1,0 +1,9 @@
+<!-- here the user can choose one of its hero's and make a new one -->
+<div class="heroDiv">
+    <p>
+       Hey, <?=$_SESSION['user_name']?>. You are logged in.<br>
+       Try to close this browser tab and open it again. Still logged in! ;)
+    </p>
+    <h1>HIER MOETEN HERO'S STAAN OFZO</h1>
+    <form><input type="submit" name="logout" value="Log out" /></form>
+</div>
