+++ application/view/templates/footer.php
@@ -1,0 +1,2 @@
+    </main>
+</body>
