+++ application/view/templates/header.php
@@ -1,0 +1,6 @@
+<!DOCTYPE html>
+<head>
+    <link type="style.css" href="" />
+</head>
+<body>
+    <main class="overall">
