+++ public/index.php
@@ -1,22 +1,30 @@
 <?php
 
-    // checking for minimum PHP version
+    /**
+    */
+    define('ROOT', dirname(__DIR__) . DIRECTORY_SEPARATOR);
+    define('APP', ROOT . 'application' . DIRECTORY_SEPARATOR);
+
+    /**
+    */
     if (version_compare(PHP_VERSION, '5.3.7', '<')) {
-        exit("Sorry, Simple PHP Login does not run on a PHP version smaller than 5.3.7 !");
+        exit('Sorry, Simple PHP Login does not run on a PHP version smaller than 5.3.7 !');
     } else if (version_compare(PHP_VERSION, '5.5.0', '<')) {
-        require_once("libraries/password_compatibility_library.php");
+        require_once(APP . 'libraries/password_compatibility_library.php');
     }
 
-    // includes the php files
-    require_once("../application/config/config.php");
-    require_once("../application/core/application.php");
+    /**
+    */
+    require_once(APP . 'config/config.php');
+    require_once(APP . 'core/database.php');
+    require_once(APP . 'core/login.php');
+    require_once(APP . 'core/register.php');
+    require_once(APP . 'core/application.php');
 
-    // creates the application object
-    $APP = new Application();
+    /**
+    */
+    $application = new Application();
+    $application->render();
+    $data = $application->getRequiredGameData();
 
-    // load the game or the login page
-    if ($APP->isUserLoggedIn() == true) {
-        include("../application/view/hero.php");
-    } else {
-        include("../application/view/login.php");
-    }
+    var_dump($data);
